+++ sketch-compose-core/src/main/kotlin/com/github/panpf/sketch/compose/SubcomposeAsyncImage.kt
@@ -63,7 +63,7 @@ fun SubcomposeAsyncImage(
     alpha: Float = DefaultAlpha,
     colorFilter: ColorFilter? = null,
     filterQuality: FilterQuality = DefaultFilterQuality,
-    noClipContent: Boolean = false,
+    clipToBounds: Boolean = true,
 ) {
     val request = DisplayRequest(LocalContext.current, imageUri)
     SubcomposeAsyncImage(
@@ -77,7 +77,7 @@ fun SubcomposeAsyncImage(
         alpha = alpha,
         colorFilter = colorFilter,
         filterQuality = filterQuality,
-        content = contentOf(loading, success, error, noClipContent),
+        content = contentOf(loading, success, error, clipToBounds),
     )
 }
 
@@ -178,7 +178,7 @@ fun SubcomposeAsyncImage(
     alpha: Float = DefaultAlpha,
     colorFilter: ColorFilter? = null,
     filterQuality: FilterQuality = DefaultFilterQuality,
-    noClipContent: Boolean = false,
+    clipToBounds: Boolean = true,
 ) = SubcomposeAsyncImage(
     request = request,
     contentDescription = contentDescription,
@@ -190,7 +190,7 @@ fun SubcomposeAsyncImage(
     alpha = alpha,
     colorFilter = colorFilter,
     filterQuality = filterQuality,
-    content = contentOf(loading, success, error, noClipContent),
+    content = contentOf(loading, success, error, clipToBounds),
 )
 
 /**
@@ -323,7 +323,7 @@ fun SubcomposeAsyncImageScope.SubcomposeAsyncImageContent(
     contentScale: ContentScale = this.contentScale,
     alpha: Float = this.alpha,
     colorFilter: ColorFilter? = this.colorFilter,
-    noClipContent: Boolean = false,
+    clipToBounds: Boolean = true,
 ) = Content(
     modifier = modifier,
     painter = painter,
@@ -332,7 +332,7 @@ fun SubcomposeAsyncImageScope.SubcomposeAsyncImageContent(
     contentScale = contentScale,
     alpha = alpha,
     colorFilter = colorFilter,
-    noClipContent = noClipContent,
+    clipToBounds = clipToBounds,
 )
 
 @Stable
@@ -340,7 +340,7 @@ private fun contentOf(
     loading: @Composable (SubcomposeAsyncImageScope.(State.Loading) -> Unit)?,
     success: @Composable (SubcomposeAsyncImageScope.(State.Success) -> Unit)?,
     error: @Composable (SubcomposeAsyncImageScope.(State.Error) -> Unit)?,
-    noClipContent: Boolean = false,
+    clipToBounds: Boolean = true,
 ): @Composable SubcomposeAsyncImageScope.() -> Unit {
     return if (loading != null || success != null || error != null) {
         {
@@ -351,10 +351,10 @@ private fun contentOf(
                 is State.Error -> if (error != null) error(state).also { draw = false }
                 is State.Empty -> {} // Skipped if rendering on the main thread.
             }
-            if (draw) SubcomposeAsyncImageContent(noClipContent = noClipContent)
+            if (draw) SubcomposeAsyncImageContent(clipToBounds = clipToBounds)
         }
     } else {
-        { SubcomposeAsyncImageContent(noClipContent = noClipContent) }
+        { SubcomposeAsyncImageContent(clipToBounds = clipToBounds) }
     }
 }